Update files for the HPSS (which replaced the MSS) for mass storage.  also remove project numbers, add headers and copyrights, etc.
git-svn-id: https://subversion.ucar.edu/DAReS/DART/branches/cam-update@4919 acad1b26-7d42-44b9-9cec-aac8293558f5

+++ full_experiment/hpss2restart.csh
@@ -1,6 +1,6 @@
 #!/bin/csh
 #
-# DART software - Copyright © 2004 - 2010 UCAR. This open source software is
+# DART software - Copyright 2004 - 2011 UCAR. This open source software is
 # provided by UCAR, "as is", without charge, subject to all terms of use at
 # http://www.image.ucar.edu/DAReS/DART/DART_download
 #
@@ -18,20 +18,20 @@ echo $LD_LIBRARY_PATH
 
 if ($#argv < 3) then
    echo "usage; from case/experiment/obs_seq directory; "
-   echo "       ms2restart MSpathname(no mss:) batch1 batchn uncompress(optnl)"
+   echo "       hpss2restart.csh HPSSpathname(no mss:) batch1 batchn uncompress(optnl)"
    echo "       untars ensemble members and puts them where they belong"
    echo "       Uncompresses caminput and clminput files, if told to."
    exit
 endif 
 
-set ms_root = $1
-set ms_dir = mss:$ms_root
+set hpss_root = $1
+set hpss_dir = $hpss_root
 
-msls -l $ms_root
+hsi -P ls -l $hpss_root
 if ($status == 0) then
-   echo "files will be read from $ms_root/batch#"
+   echo "files will be read from $hpss_root/batch#"
 else
-   echo "$ms_root does not exist.  Check name and try again"
+   echo "$hpss_root does not exist.  Check name and try again"
    exit
 endif
 
@@ -43,29 +43,49 @@ set batch = $2
 
 while($batch <= $batchn)
 
-   echo "msrcping ${ms_dir}/batch${batch}"
+   echo "getting ${hpss_dir}/batch${batch} to batch$batch"
    if ($uncomp == true) then
-      msrcp ${ms_dir}/batch${batch}.cmp batch$batch
+      hsi get batch$batch : ${hpss_dir}/batch${batch}.cmp
    else
-      msrcp ${ms_dir}/batch$batch .
+      hsi get batch$batch : ${hpss_dir}/batch$batch
    endif
 
-#  This makes CAM CLM DART directories, as necessary, to place its files.
+   # This makes CAM CLM DART directories, as necessary, to place its files.
    echo "untarring"
-   tar x -f batch$batch
-   rm batch$batch
+   tar -x -f batch$batch
+   if ($status == 0) then
+      rm batch$batch
+   else
+      echo "Could not untar batch$batch"
+   endif
    @ batch++
 end
 
 # Get filter ics when all were written to one file.
 # Also get assim_tools_ics, if it exists.
 # DART may exist from untarring, above.
-mkdir DART
-if (! -e DART/filter_ic.0001) then
-   echo "msrcping single filter_ic and/or inflation ic file(s)"
-   msls ${ms_root}/DART/ic_files.tar
+if (! -d DART) then
+   mkdir DART
+endif
+if (! -f DART/filter_ic.0001) then
+   echo DART has no filter_ic.0001
+   ls DART
+   hsi -P ls ${hpss_root}/DART/filter_ic 
+   if ($status == 0) 
+      echo "Getting single filter_ic"
+      cd DART
+      hsi get filter_ic : ${hpss_dir}/DART/filter_ic 
+      cd ..
+   else
+      echo 'no filter_ic (nor filter_ic.0001) found'
+   endif 
+
+   hsi -P ls ${hpss_root}/DART/assim_tools_ics
    if ($status == 0) then
-      msrcp ${ms_dir}/DART/ic_files.tar DART
+      cd DART
+      hsi get assim_tools_ics : ${hpss_dir}/DART/assim_tools_ics  
+   else
+      echo 'no assim_tools_ics found'
    endif
 endif
 
@@ -74,7 +94,7 @@ if ($uncomp == true) then
    gunzip -r CLM
 endif 
 
-echo "msrcp done; check local files "
+echo "Finished fetching files; check local files "
 
 exit 0
 
@@ -82,4 +102,5 @@ exit 0
 # $URL$
 # $Revision$
 # $Date$
+
